requirements.txt
# Core dependencies
streamlit>=1.28.0
PyPDF2>=3.0.0
python-docx>=1.0.0
nltk>=3.8.0
scikit-learn>=1.3.0
pdfplumber>=0.10.0

# For API integrations
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
